framework/requirements.txt
# TestZen - No-Code Test Automation Framework
# Required Python packages

# Core Appium and Selenium
Appium-Python-Client==3.1.1
selenium==4.15.0

# Excel file handling
pandas==2.1.0
openpyxl==3.1.2
xlrd==2.0.1

# Configuration files
PyYAML==6.0.1

# Logging and utilities
colorlog==6.7.0

# Web driver management
webdriver-manager==4.0.1

# Optional: For advanced features
requests==2.31.0
Pillow==10.0.0
jinja2==3.1.2